+++ src/extension_core/extension.ts
@@ -78,6 +78,7 @@ export async function activate(context: vscode.ExtensionContext): Promise<Vaporv
   const signalSelectEvent = WaveformViewerProvider.signalSelectEventEmitter.event;
   const addVariableEvent = WaveformViewerProvider.addVariableEventEmitter.event;
   const removeVariableEvent = WaveformViewerProvider.removeVariableEventEmitter.event;
+  const valueLinkEvent = WaveformViewerProvider.valueLinkEventEmitter.event;
   const externalDropEvent = WaveformViewerProvider.externalDropEventEmitter.event;
 
   // Register commands (there are a lot of commands, so we register them in a separate file for cleanliness)
@@ -92,6 +93,7 @@ export async function activate(context: vscode.ExtensionContext): Promise<Vaporv
     onDidAddVariable: addVariableEvent,
     onDidRemoveVariable: removeVariableEvent,
     onDidDropInWaveformViewer: externalDropEvent,
+    onDidClickSignalValueLink: valueLinkEvent,
 
     // Commands
     async openFile(args: OpenFileArgs) {